Fix RadioButton a11y
